Merge `start-bind` and `finish-bind` into a single `bind`.
Similarly, merge `start-listen` and `finish-listen` into a single
`listen`. The underlying `bind` and `listen` operations don't do any
network I/O; they just modify the local state of a socket. And they
aren't considered blocking in common host OS implementations. So it
simplifies both host and guest code to combine these into single
functions that just do the work directly.

After this change, the only `start-`/`finish-` pair remaining is
`start-connect`/`finish-connect`. I believe that makes sense to keep
as-is because it pretty closely corresponds to how POSIX `connect`
behaves on non-blocking sockets.

With this change, I also think it makes sense to remove the
`concurrency-conflict` error code, and just use `already-connected`
for the cases where an operation doesn't work if a connection is
either established or in the process of being established.